AUTH_TEST/copilot/copilot1/backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
pydantic==2.5.1
python-jose==3.3.0
passlib==1.7.4
bcrypt==4.0.1
python-multipart==0.0.6
aiosqlite==0.19.0
